+++ inc/wd-register-post-type.php
@@ -13,15 +13,15 @@ if ( ! defined( 'ABSPATH' ) ) {
 }
 
 $admin_skin = get_user_option('admin_color');
-			
+
 if ( $admin_skin == 'light' ) {
-	
+
 	$icon_url = WD_URI . '/assets/img/admin/vynil-dark.png';
 } else {
 
 	$icon_url = WD_URI . '/assets/img/admin/vynil.png';
 }
-	
+
 $labels = array(
 	'name' => esc_html__( 'Releases', '%TEXTDOMAIN%' ),
 	'singular_name' => esc_html__( 'Release', '%TEXTDOMAIN%' ),