+++ pack/wolf-discography/inc/admin/class-wd-options.php
@@ -6,7 +6,7 @@
  * @author WolfThemes
  * @category Admin
  * @package WolfDiscography/Admin
- * @version 1.4.0
+ * @version 1.4.7
  */
 
 if ( ! defined( 'ABSPATH' ) ) {
@@ -70,9 +70,9 @@ class WD_Options {
 		register_setting( 'wolf-release-settings', 'wolf_release_settings', array( $this, 'settings_validate' ) );
 		add_settings_section( 'wolf-release-settings', '', array( $this, 'section_intro' ), 'wolf-release-settings' );
 		add_settings_field( 'page_id', esc_html__( 'Discography Page', 'wolf-discography' ), array( $this, 'setting_page_id' ), 'wolf-release-settings', 'wolf-release-settings' );
-		add_settings_field( 'use_band_tax', esc_html__( 'Link Band Name to category page', 'wolf-discography' ), array( $this, 'setting_use_band_tax' ), 'wolf-release-settings', 'wolf-release-settings' );
-		add_settings_field( 'use_label_tax', esc_html__( 'Link Label Name to category page', 'wolf-discography' ), array( $this, 'setting_use_label_tax' ), 'wolf-release-settings', 'wolf-release-settings' );
-		add_settings_field( 'display_format', esc_html__( 'Display format (like CD, digital download etc...)', 'wolf-discography' ), array( $this, 'setting_display_format' ), 'wolf-release-settings', 'wolf-release-settings' );
+		add_settings_field( 'use_band_tax', esc_html__( 'Link Artist Name', 'wolf-discography' ), array( $this, 'setting_use_band_tax' ), 'wolf-release-settings', 'wolf-release-settings' );
+		add_settings_field( 'use_label_tax', esc_html__( 'Link Label Name', 'wolf-discography' ), array( $this, 'setting_use_label_tax' ), 'wolf-release-settings', 'wolf-release-settings', array( 'class' => 'wolf-discography-settings-link-label' ) );
+		add_settings_field( 'display_format', esc_html__( 'Display format (like CD, digital download etc...)', 'wolf-discography' ), array( $this, 'setting_display_format' ), 'wolf-release-settings', 'wolf-release-settings', array( 'class' => 'wolf-discography-settings-display-format' ) );
 	}
 
 	/**
@@ -112,12 +112,22 @@ class WD_Options {
 	 * @return string
 	 */
 	public function setting_page_id() {
+		$page_option = array( '' => esc_html__( '- Disabled -', 'wolf-discography' ) );
 		$pages = get_pages();
+
+		foreach ( $pages as $page ) {
+
+			if ( get_post_field( 'post_parent', $page->ID ) ) {
+				$page_option[ absint( $page->ID ) ] = '&nbsp;&nbsp;&nbsp; ' . sanitize_text_field( $page->post_title );
+			} else {
+				$page_option[ absint( $page->ID ) ] = sanitize_text_field( $page->post_title );
+			}
+		}
 		?>
 		<select name="wolf_release_settings[page_id]">
 			<option value="-1"><?php esc_html_e( 'Select a page...', 'wolf-discography' ); ?></option>
-			<?php foreach ( $pages as $page ) : ?>
-				<option value="<?php echo absint( $page->ID ); ?>" <?php selected( absint( $page->ID ), get_option( '_wolf_discography_page_id' ) ); ?>><?php echo sanitize_text_field( $page->post_title ); ?></option>
+			<?php foreach ( $page_option as $k => $v ) : ?>
+				<option value="<?php echo absint( $k ); ?>" <?php selected( absint( $k ), get_option( '_wolf_discography_page_id' ) ); ?>><?php echo sanitize_text_field( $v ); ?></option>
 			<?php endforeach; ?>
 		</select>
 		<?php