+++ pack/wolf-discography/templates/content-release.php
@@ -4,7 +4,7 @@
  *
  * @author WolfThemes
  * @package WolfDiscography/Templates
- * @version 1.4.0
+ * @version 1.4.7
  * @since 1.0.2
  */
 ?>