+++ pack/wolf-discography/uninstall.php
@@ -7,7 +7,7 @@
  * @author WolfThemes
  * @category Core
  * @package WolfDiscography/Uninstaller
- * @version 1.4.0
+ * @version 1.4.7
  */
 
 if ( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {