+++ pack/wolf-discography/wolf-discography.php
@@ -3,11 +3,11 @@
  * Plugin Name: Wolf Discography
  * Plugin URI: https://wolfthemes.com/plugin/wolf-discography/
  * Description: A discography plugin to display your releases
- * Version: 1.4.0
+ * Version: 1.4.7
  * Author: WolfThemes
  * Author URI: https://wolfthemes.com
  * Requires at least: 4.4.1
- * Tested up to: 4.7.2
+ * Tested up to: 4.9.5
  *
  * Text Domain: wolf-discography
  * Domain Path: /languages/
@@ -20,7 +20,7 @@
  * Verified customers however, who have purchased a premium theme
  * at https://themeforest.net/user/Wolf-Themes/portfolio?ref=Wolf-Themes
  * will have access to support for this plugin in the forums
- * https://help.wolfthemes.com/
+ * https://wolfthemes.ticksy.com/
  *
  * Copyright (C) 2013 Constantin Saguin
  * This WordPress Plugin is a free software: you can redistribute it and/or modify
@@ -46,7 +46,7 @@ if ( ! class_exists( 'Wolf_Discography' ) ) {
 	 * Contains the main functions for Wolf_Discography
 	 *
 	 * @class Wolf_Discography
-	 * @version 1.4.0
+	 * @version 1.4.7
 	 * @since 1.0.0
 	 * @package WolfDiscography
 	 * @author WolfThemes
@@ -61,7 +61,7 @@ if ( ! class_exists( 'Wolf_Discography' ) ) {
 		/**
 		 * @var string
 		 */
-		public $version = '1.4.0';
+		public $version = '1.4.7';
 
 		/**
 		 * @var Wolf Discography The single instance of the class
@@ -97,20 +97,6 @@ if ( ! class_exists( 'Wolf_Discography' ) ) {
 				self::$_instance = new self();
 			}
 			return self::$_instance;
-		}
-
-		/**
-		 * Cloning is forbidden.
-		 */
-		public function __clone() {
-			_doing_it_wrong( __FUNCTION__, esc_html__( 'Cheatin&#8217; huh?', 'wolf-discography' ), '1.0' );
-		}
-
-		/**
-		 * Unserializing instances of this class is forbidden.
-		 */
-		public function __wakeup() {
-			_doing_it_wrong( __FUNCTION__, esc_html__( 'Cheatin&#8217; huh?', 'wolf-discography' ), '1.0' );
 		}
 
 		/**
@@ -162,7 +148,7 @@ if ( ! class_exists( 'Wolf_Discography' ) ) {
 		 * Activation function
 		 */
 		public function activate() {
-			
+
 			add_option( '_wolf_discography_needs_page', true );
 
 			if ( ! get_option( '_wolf_discography_flush_rewrite_rules_flag' ) ) {
@@ -184,7 +170,7 @@ if ( ! class_exists( 'Wolf_Discography' ) ) {
 		 * Define WD Constants
 		 */
 		private function define_constants() {
-			
+
 			$constants = array(
 				'WD_DEV' => false,
 				'WD_DIR' => $this->plugin_path(),