+++ templates/single-release.php
@@ -55,7 +55,6 @@
 					 * Tracklists
 					 */
 					wd_release_tracklist();
-
 				?>
 			</div>